+++ Examples/Modules/Catalog/DDDToolkit.Examples.Catalog.Migrations.SqlServer/Migrations/20260925111440_OutboxNextAttemptAt.Designer.cs
@@ -5,16 +5,19 @@ using DDDToolkit.Examples.Catalog.Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
 
-namespace DDDToolkit.Examples.Catalog.Migrations.SqlServer
+namespace DDDToolkit.Examples.Catalog.Migrations.SqlServer.Migrations
 {
     [DbContext(typeof(CatalogContext))]
-    partial class CatalogContextModelSnapshot : ModelSnapshot
+    [Migration("20260925111440_OutboxNextAttemptAt")]
+    partial class OutboxNextAttemptAt
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder