+++ Examples/Modules/Payments/DDDToolkit.Examples.Payments/Infrastructure/Persistence/Migrations/20260925111355_OutboxNextAttemptAt.Designer.cs
@@ -4,17 +4,20 @@ using System.Collections.Generic;
 using DDDToolkit.Examples.Payments.Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
 #nullable disable
 
-namespace DDDToolkit.Examples.Payments.Migrations
+namespace DDDToolkit.Examples.Payments.Infrastructure.Persistence.Migrations
 {
     [DbContext(typeof(PaymentsContext))]
-    partial class PaymentsContextModelSnapshot : ModelSnapshot
+    [Migration("20260925111355_OutboxNextAttemptAt")]
+    partial class OutboxNextAttemptAt
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder